third_party/blink/web_tests/fast/inline/inline-position-top-align-expected.txt

pre-text  after-text	Something



Tests for hitTest of element
bug 45164: REGRESSION:  Clickable area too large

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS hitResult is false
